+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## fontminify@0.0.2 (2026-03-13)
+
+- Update description and README.
+
 ## fontminify@0.0.1 (2026-03-13)
 
 - First release